requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.31.0
celery==5.3.1
django-celery-beat==2.5.0
